Move formatting tests to a specific module
